+++ src/api-script/src/main/java/org/rbernalop/apiscript/script/application/finder/by_userId/FindScriptsByUserIdQueryHandler.java
@@ -1,4 +1,4 @@
-package org.rbernalop.apiscript.script.application.finder.userId;
+package org.rbernalop.apiscript.script.application.finder.by_userId;
 
 import lombok.extern.slf4j.Slf4j;
 import org.rbernalop.apiscript.script.application.finder.ScriptFinder;
@@ -6,8 +6,8 @@ import org.rbernalop.apiscript.script.domain.repository.ScriptRepository;
 import org.rbernalop.apiscript.script.domain.service.DomainScriptFinder;
 import org.rbernalop.apiscript.script.domain.value_object.PaginationRequest;
 import org.rbernalop.apiscript.script.domain.value_object.OwnerUsername;
-import org.rbernalop.apiscript.shared.application.script.find.FindScriptsByUserIdQuery;
-import org.rbernalop.apiscript.shared.application.script.find.FindScriptsByUserIdResponse;
+import org.rbernalop.apiscript.shared.application.script.find.by_userid.FindScriptsByUserIdQuery;
+import org.rbernalop.apiscript.shared.application.script.find.by_userid.FindScriptsByUserIdResponse;
 import org.rbernalop.shared.domain.Service;
 import org.rbernalop.shared.domain.bus.query.QueryBus;
 import org.rbernalop.shared.domain.bus.query.QueryHandler;
