+++ src/api-script/src/main/java/org/rbernalop/apiscript/script/infrastructure/controller/ScriptGetByUserIdController.java
@@ -3,8 +3,8 @@ package org.rbernalop.apiscript.script.infrastructure.controller;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.rbernalop.apiscript.shared.application.script.find.FindScriptsByUserIdQuery;
-import org.rbernalop.apiscript.shared.application.script.find.FindScriptsByUserIdResponse;
+import org.rbernalop.apiscript.shared.application.script.find.by_userid.FindScriptsByUserIdQuery;
+import org.rbernalop.apiscript.shared.application.script.find.by_userid.FindScriptsByUserIdResponse;
 import org.rbernalop.apiscript.shared.application.script.find.ScriptResponse;
 import org.rbernalop.shared.domain.bus.command.CommandBus;
 import org.rbernalop.shared.domain.bus.query.QueryBus;
