+++ src/api-script/src/main/java/org/rbernalop/apiscript/shared/application/script/find/by_userid/FindScriptsByUserIdResponse.java
@@ -1,9 +1,10 @@
-package org.rbernalop.apiscript.shared.application.script.find;
+package org.rbernalop.apiscript.shared.application.script.find.by_userid;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 import org.rbernalop.apiscript.script.domain.aggregate.Script;
+import org.rbernalop.apiscript.shared.application.script.find.ScriptResponse;
 import org.rbernalop.shared.domain.bus.query.Response;
 
 import java.util.List;
