+++ src/api-script/src/test/java/org/rbernalop/apiscript/script/application/finder/by_userId/FindScriptsByOwnerUsernameQueryHandlerTest.java
@@ -1,4 +1,4 @@
-package org.rbernalop.apiscript.script.application.finder.userId;
+package org.rbernalop.apiscript.script.application.finder.by_userId;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -7,7 +7,7 @@ import org.rbernalop.apiscript.script.domain.aggregate.Script;
 import org.rbernalop.apiscript.script.domain.aggregate.ScriptMother;
 import org.rbernalop.apiscript.script.domain.repository.ScriptRepository;
 import org.rbernalop.apiscript.script.domain.value_object.OwnerUsername;
-import org.rbernalop.apiscript.shared.application.script.find.FindScriptsByUserIdQuery;
+import org.rbernalop.apiscript.shared.application.script.find.by_userid.FindScriptsByUserIdQuery;
 import org.rbernalop.apiscript.shared.application.script.find.FindScriptsByUserIdQueryMother;
 import org.rbernalop.apiscript.shared.domain.exception.NegativeException;
 import org.rbernalop.shared.domain.bus.query.QueryBus;
